requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
websockets==12.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
firebase-admin==6.3.0
slowapi==0.1.9
aiofiles==23.2.1
